Rewrite/restructure project to be easier to use.
These changes have not been tested on data and there will be bugs.
These changes have to be commited to be tested on Zemenar.

Changes:
1. Use new pip installed auth_ident imports

2. GenericExecute
	This class is the super class for the training/validation or any
	other tasks we will need to do. This contains a lot of the shared
	logic between tasks.

	Some of the shared logic includes:

	* The argparser
	* Geting getting the correct directory for each expirement and combination
	* Loads param dict and creates neccessary combinations
	* The primary model loop
	* Creating the logger and some basic logging
	* Set up a structure for loading, updating, and saving hyperparameter
	results

	This class is set up in a way to easily overwrite each method for custom
	functions which can varry widley.

3. auth_ident/train_contrastive.py contains the train script for
training the contrastive model.

4. auth_ident/train_secondary.py contains the train script for
the secondary model. This replaces the outer_model.py script that
was not very reusable and didn't save or log models or data.

5. new format for param_dict.json because of the secondary trainer.
Should contains two dictionaries 'contrastive' and `secondary` where
params for the contrastive training and secondary model training
respectively.

6. GenericSecondaryClassifier super class for all of the secondary classifiers.
A wrapper was created for all of the SK-Learn classifiers we were using.

7. Reorganized location of some files.