Add end to end secondary classifier.
This classifier has to be dealt with differently than other secondary
classifiers because we cannot just pass the embeddings. We need to give
it the loaded model and the raw data.

This means the embedding util function was changed and broken up a bit.
It is more module to allow parts of it to be called directly and get the
data and model separately from the embeddings.

The secondary classifier also has code to build the CNN from previous
author ident. work. To use it you have to change the function being called
in the training function.

There is likely a much cleaner way to do this implementation, but it is
probably not worth refactoring the code at this point, especially with our
new open set focus.